State of Georgia Website Scraper/requirements.txt
attrs==23.2.0
boto3==1.34.34
botocore==1.34.34
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
et-xmlfile==1.1.0
h11==0.14.0
idna==3.6
jmespath==1.0.1
lxml==5.1.0
numpy==1.26.3
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pycparser==2.21
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
s3transfer==0.10.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
webdriver-manager==4.0.1
wsproto==1.2.0
